+++ invenio_s3/ext.py
@@ -9,7 +9,6 @@
 
 import warnings
 
-import boto3
 from flask import current_app
 from werkzeug.utils import cached_property
 
@@ -49,6 +48,7 @@ class InvenioS3(object):
                 DeprecationWarning,
             )
 
+        s3_config_extra = current_app.config.get("S3_CONFIG_EXTRA", {})
         info = dict(
             key=current_app.config.get("S3_ACCESS_KEY_ID", ""),
             secret=current_app.config.get("S3_SECRET_ACCESS_KEY", ""),
@@ -60,6 +60,7 @@ class InvenioS3(object):
                 "signature_version": current_app.config.get(
                     "S3_SIGNATURE_VERSION", "s3v4"
                 ),
+                **s3_config_extra,
             },
         )